requirements.txt
openai==1.7.0
python-dotenv==1.0.0
opencv-python==4.9.0.80
requests==2.31.0
Pillow==10.2.0